pyproject.toml
[project]
name = "persona"
version = "0.1.0"
description = "Intelligent User Memory System"
packages = [
    { include = "persona", from = "." },
]
authors = [
    { name = "Utkarsh Saxena", email = "utkarsh@buildpersona.ai" },
]
homepage = "https://github.com/saxenauts/persona"
repository = "https://github.com/saxenauts/persona"

[tool.poetry]
name = "persona"
version = "0.1.0"
description = "A digital user memory and identity system for personalization"
authors = ["Utkarsh Saxena"]
packages = [
    { include = "persona", from = "." },
]
license = "MIT"
homepage = "https://github.com/saxenauts/persona"
repository = "https://github.com/saxenauts/persona"

[tool.poetry.dependencies]
python = "~3.12.0"
fastapi = "^0.111.0"
neo4j = "^5.20.0"
openai = "^1.30.3"
pydantic-settings = "^2.2.1"
tiktoken = "^0.8.0"
sqlalchemy = "^2.0.36"
aiosqlite = "^0.20.0"
requests = "^2.32.3"
anthropic = "^0.40.0"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.0"
mem0ai = "^1.0.1"
graphiti-core = "^0.24.3"
langchain-openai = "^1.1.3"
azure-identity = "^1.25.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
pytest-asyncio = "^0.21.0"
httpx = "^0.24.1"
pycurl = "^7.45.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"